Assignment1/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fonttools==4.61.1
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
